Refactor!: Added Vectors type

+++ bot/core/Controller.go
@@ -1,6 +1,6 @@
 package core
 
 type Controller struct {
-	Left, Right, Forward, Back float32
+	Left, Right, Forward, Back float64
 	Jump, Sneak, Sprint, Drop  bool
 }
